Calculations Rewrite + Removed Automatic Installation
The number of rows in the textarea is now calculated by dividing the textarea's scrollHeight by it's text line height, instead of splitting the textarea's value on each instance of a newline character. The old implementation didn't run well when typing in a textarea with a large text value; 1000+ lines usually started to cause input delay and freezing frames.

This commit also removes the ability to install the plugin with one JavaScript file in favor of being more controllable by the person choosing to work with it. The primary way to install the plugin is now by linking to the style and script files individually, as well as wrapping the designated textareas to be numbered with a container on the client side instead of using the data-lined-automatic attribute which would wrap the textarea with a span element automatically. This is now meant to be handled by the developer rather than the plugin!