Calculations Re-Rewrite
The last update appeared to actually make the issue worse than it was before the rewrite, so I went back to the original implementation with a few fixes that should help with a few more issues that could occur. Typing performance in textareas with 1000+ lines of text is still a sub-par experience for what I hope to achieve, so I hope I will be able to find a better performing method than how it currently functions.