Web Component Rewrite
I reworked the plugin from the ground up to use the Web Components API! Instead of creating a numbered textarea with <div> and <textarea> elements, you only have to add a <num-text> element, and all of the heavy lifting from the plugin will be hidden away in it's Shadow DOM. Also, you only have to include the plugin's script tag in your document now, not the stylesheet. The plugin will add the styles to the <num-text> element's Shadow DOM directly using a <link> tag.