Pointer Events Detection Issue
I forgot that the "pointerdown" event fires exactly like "touchdown" on touchscreens, so I incorrectly changed the "mousedown" event listeners to "pointerdown" ones. This is now back to how it was before. With a mouse, you just have to click down for the NumText element to get focused, and on touch devices, you have to fully tap on the element for it to focus. This is the touch equivalent of "click" on touch devices.