The Next Rewrite
In a similar vein to my recent work on Menu Drop, I am working on some nicer reworks for the components, now that I've had some more experience with Web Components since originally working on these!

I had some time away from home today to code (kind of a refresher for more inspiration), and I was actually going to try out using CodeMirror 6, since that's one of the libraries that has inspired Num Text, but I didn't have Wi-Fi there to clone the repo! Funny issue hehe. I decided to just start reworking an extra simple STE for fun, and I got a good amount of Num Text redone from scratch! Some parts are fairly similar to the current codebase, but I made a lot of the similar changes that I did to Menu Drop recently, which I've also wanted to do with Num Text. This commit is just a few changes to align the future work with that new design.

Alright, this next part was actually after that previous CodeMirror sentence, so the context is why I am working on my own editor, rather than using CodeMirror or another editor instead.

I like the feature set that it has, as well as the mobile support, but I'm not too keen on the complexity of everything. It's nothing against CodeMirror by any means at all, it's just my weird self that likes to have things nice and encapulated. I haven't needed to use a bundler yet for my projects, and I don't want to start using one just for the sake of a code editor library. This seems to be a disconnect for many of the JS code editors out there. While they are designed for the web, they don't appearently provide HTML support to just use them in your markup, out of the box. I really like CodeMirror, Monaco Editor, and Ace too, but they aren't as simple of a process to use as I'd hope them to be. Rather than fretting about their effort and hard work to make such powerful libraries, I think it it worthwhile to try and solve my own problem myself, since I have the power to make it however I need it to work :)

Coincidentally, I started going down the Medium rabbit-hole today, and found this article, written by the author of Babel.
https://medium.com/@sebmck/2015-in-review-51ac7035e272

Not to say that my projects do, or will have the depth that his have had, but I found his story remarkably familiar to the one I've had over the last few years, and it was almost frightening! I really find it nice to try and solve my own challenges too, even if I already have an option to solve it using the more mainstream way. I have felt this way with a good amount of my coding ventures. To name a few, using PWAs over native apps or Electron, using Web Components or vanilla JS rather than a web framework, or making my own frickin' text editor! XD
It was almost comical to me that I frequently find myself doing this, and I do it with a lot of other things in my non-coding life too. From the unheard-of music nobody has heard of yet, or running as many miles as I am old, which is a goal I am holding up every year, until it is too much to bear. Who does that?? Not sure, guess that's why I want to do it :D

Sorry to wander with this again, I just have to say that I have really liked adding on to my changelogs like this. It is actually very relieving to write out my thoughts like this. *Incoming thought while writing that last sentence* Rather than making a blog that I update when I want to write, what if I make my blog just based on my commit changelogs? It could fetch all of them from my Git repos, and display them as if they were articles themselves. That sounds epic! It would likely be best to be SSR-based. Maybe cached on the server too, and I can just update the commit messages database once I upload a few commits. Then the server can do the rest of the work for me by rendering them out as articles (Do the work for me, in terms of I write a crazy server that does all of this magic, then it is hassle free XD).

Alright, and to compare to where I'd eventually like to go with Num Text, checkout the great work over at CodeMirror here!
https://codemirror.net/6/

The next few updates will be removing syntax highlighting for the time being, as I want to look into doing it a nicer way than I had before. Prism definitely works perfect, but I think I could make use of it better by making things more responsive. As of now, every single time you type into the editor, all of the highlighting is re-rendered. Definitely not efficient! I don't really have a real idea on how the other libraries manage to make it more dynamic, so I plan to look more into that.