Experimental TypeScript!
Alright, so I think I really am a fan of TypeScript now, who'da figure XD
Decided to do some refactors, including bringing everything over to TypeScript!
This adds Constructable Stylesheet, ES Module, and code completion support!
This is semi-experimental, as I'm not totally sure if I'm going 100% with TypeScript yet. The library changes are closely for certain though, TypeScript or not.
Can't wait to see how this goes!

Been working on this the last 3 days or so, and there have been a lot of thing to look up:
https://github.com/43081j/shiki-element
https://github.com/calebdwilliams/construct-style-sheets#example-commands
https://caniuse.com/mdn-api_document_adoptedstylesheets
https://stackoverflow.com/questions/58120940/javascript-create-array-of-objects-of-length-n
https://stackoverflow.com/questions/6795585/remove-multiple-elements-with-same-name-using-removechild
https://developer.mozilla.org/en-US/docs/Web/API/Element/replaceChildren
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/slice
https://developer.mozilla.org/en-US/docs/Web/API/CSSStyleSheet/replace
https://stackoverflow.com/questions/23072286/where-is-the-syntax-for-typescript-comments-documented
https://stackoverflow.com/questions/28900077/why-is-event-target-not-element-in-typescript
https://stackoverflow.com/questions/59614647/property-focus-does-not-exist-on-element
https://stackoverflow.com/questions/66892113/remove-multiple-specific-childnodes-in-javascript
https://stackoverflow.com/questions/72051725/does-element-replacechildren-function-work-with-duplicate-values-or-what-is-s