Old Docs Removal
Removed the documentation I wrote a while ago. I think I did a great job at it actually! I'm gonna rework things though, so I don't want to update it along the way, I'd rather start from scratch after getting things reworked again.

Next on the list of to-dos, I've been wanting to make `<num-text>` API-compatible with the vanilla `<textarea>` element pretty much since the beginning of the component. I realized, now that I'm using TypeScript, it's actually not too hard to accomplish that! I can simply add `implements HTMLTextAreaElement` to my `class NumText` declaration, and it will show any errors for inconsistencies with properties and such. Super awesome! Without TypeScript, I was going to spend a ton of time on MDN looking over the documentation to make sure that I covered everything that `<textarea>` elements do, which would have been a lot of work to try and keep track of. With having the types available to me right in the editor, I can ensure that everything is the same! Along with a general re-lookover of the codebase, now that I have TypeScript in my toolbox, I will rework the current codebase, and add the `<textarea>` crossover properties, since they essentially just need to mirror down to the internal `<textarea>` that Num Text uses.

Part of that, I just remembered that I have to mirror those same attributes as properties on the Num Text element itself too, which also allows for styling hooks from the outside. I love it!!!