MyEditor Rework
Did a little re-learning of building the component from the start, again. I did it with the new module script way of doing it, and with Constructable Stylesheets, both embedded right into the HTML demo! I also enabled type checking for the JS, which actually works fairly well with `@ts-check`, right inside of the`<script>` tag. It runs into some library errors, I think because it doesn't recognize the Module and Target settings implicitly there. While it's inside of a `type="module"` script, it didn't like the use of the dynamic `await import()` function. Along similar lines, there's no way to enable strict type checking for scripts in that spot, so I also had a harder time looking for `null` and `undefined` values, as it didn't show those in most instances.

Yeah, so the things that I learned from that test demo, I'm bringing up to the main repo, to see how they fit in here :)