Line Indices
These are great examples where seeing TypeScript's definitions for return types and interfaces really do help a lot! You can see how the content you're working with is structured, and adjust your code to better align working with the data!

I essentially took out a layer here, as now I see that what I had abstracted into a separate method, isn't completely necessary here. The `getStringIndices()` method was going to be used by STE's JSON Formatter widget, to try and extract where parsing errors are inside of the JSON content (the old `indexi()` function over in STE).

But, two things now; One, it's not actually that hard to calculate this out, now that I know how it works; and two, it's not even a reliable application to use it there, as the code for that was flawed in the first place, haha.

It relied on a browser-specific error message which happened to list the JSON parse error line offsets, and I wanted to try and use those to decode the original text content, to get a snippet out of where the issue was, to show the user that. I really like that idea, so maybe there's another way to do it, not based on the error message specifically.

Yeah, so this commit hides some of the abstractions back into the internal code, since it isn't something big enough to need to be used elsewhere, and it also refactors the remaining code to be a little more streamlined and easy on the eyes, at least more-similar structures to how I've been tending to program recently :)