Gutter Methods
Broke up the `#refreshGutter()` method into more-specific methods, which separately either add or remove line numbers as needed. It moves the "how it works" code out of the conditional, "how it will run" code.

I also renamed `#refreshGutter()` to `#renderGutter()`, as that's more in line (pun-unintended) and specific to what that method is doing.

And I reorganized how the `value` getter and setters are listed. I'm going to move away from the `document.execCommand()` calls for the `value` setter, and try setting up my own undo-redo stack instead. I want this to be a more-capable editor, and I think that will allow for more things to be possible down the line :)