Root Scrolling
This is a new attempt at trying to handle scrolling using the root element instead of syncing up the gutter alongside the textarea.

This version is actually working very well! It makes the scroll animations for everything much smoother on iOS, since here it makes use of the built in scroll system, allowing for fractional scroll positions when moving.

There are still some issues though. Moving the caret to a place covered by the gutter won't scroll the content in to view quite enough, so the caret is hidden underneath the gutter. I think a workaround for this would be to use something other than `position: sticky` on the gutter. That works perfectly on iOS though, and it's exactly what I'm looking for, so maybe a clearfix or something like that could fix it to work correctly.

Safari also has some trouble with the root `isolation: isolate` and `z-index` values on the component children, in regards to the focus ring. It continues to render parts of it inside of the scrolling container, fairly broken surprisingly.

Overall, it removes a lot of magic code that I had made for it to work using my workaround, but this is more straightforward code to make it the vanilla way, maybe just dealing with more browser box model and rendering bugs instead.