Standalone npm Setup
Preparing the legacy version of the component to be able to run independent of where it's run from. The currently active version is relying on my GitHub Pages deployment to be active, and to use that to load the component's stylesheets from. This is a problem to be able to have everything just work by itself, which is what I'm trying to modulate the package to be able to do.

This update even moreso moves the codebase over to a more standard TypeScript setup, now that I plan on packaging it for npm. No need to have the JavaScript source and the TypeScript declarations for it, as this is what will represent the final legacy version of the package itself. Moving this to npm is already going to break all of my projects that rely on it being deployed from my GitHub Pages build of it, so I might as well make it something that I can fall back to for the projects I can't or don't want to migrate up to the new Num Text. I have some random demos that use text editors that I will have to move over to using a CDN of this build, since the GitHub Pages deployment won't be accessible to run from anymore, once I move everything to npm.

I'm also doing this (mainly, to be honest) because the current production of STE is one of the projects relying on this goofed setup, and I don't want to blindly migrate to the new Num Text version before it's ready, just because this one has to go. I for certain know that they don't have feature parity, so I can't safely use the new version without something for sure breaking in STE. It also means I don't have to migrate right away, which means I can finish helping get the new Num Text up to par in features with the old one, which would be absolutely awesome.

This legacy version of Num Text is exactly on par with all of the features I wanted it to have, it just turns out that I built it without modularity in mind, it wasn't type-safe, I didn't version it (so there weren't any fallbacks if something broke in production), and I also didn't deploy it dependably. That's what the new version of Num Text is accomplishing. It's both rewriting this version to fix all of it's technical debt problems, and it's also taking the spot of this one, as I currently depend on this one being around, and to be able to use it.

I couldn't remember the phrase to describe this, so I asked ChatGPT to see if it knew what I was talking about. The one I thought I remember hearing somewhere was "Rebuilding your house while living in it". One of ChatGPT's answers was "Changing the tires on a moving car". That one is a great description for it! That's exactly what's going on here.