Small Formatting
This is the first part to the formatting I'm doing to make things easier to read. These are the smaller changes that I know won't affect too much.

Alongside these first changes, I realized that the Num Text-specific types are also added to the global scope in TypeScript (because this version isn't designed to be loaded using ESM), so I prefaced them with `NumText` to better ensure that they don't collide with other definitions in the global scope that may be similarly named.