Big Formatting
Since I most likely will come back to this code again to make the new version of Num Text feature synonymous with this one, I think going ahead and formatting everything is probably the best move in the long run. Having everything laid out like how I organize my code now will only be of help to me anyways, why not? :D

Note that this isn't refactoring the code to work better, it's simply just more readable now.

I debated changing all of the applicable `var` statements local to functions to `const` or `let` ones, but I don't want to commit to that before knowing if anything currently relies on it being mutable. Of course TypeScript gives me errors as to whether I accidentally made something immutable, but I want to at least do more testing on the existing code before changing anything to use `const`, just to make it look nicer. It does make it more straightforward to use `const` though, so I think I will at least look into it again in a commit after this one to make sure nothing goes wrong by doing that.

I was trying out the demo page here that shows a live version of a Num Text editor, and I looked into the Prism theme styles feature that I never fully took advantage of in this codebase, and it really is excellent! I bootstrapped it into a dev build of STE, and I loaded the VSC-Dark-Plus Prism theme in. It looked SPECTACULAR! Now I totally remember how cool that is. I definitely want to update this functionality to be present in the new Num Text rewrite. I was debating leaving syntax highlighing behind in favor of slimming things down a bit, but I think I crossed that bridge now that things are all type-safe and part way to being all refactored. It's gonna be much easier to manage now that I have TypeScript and Constructable Stylesheets on my side too.