API Documentation!
Improving the documentation of the API! I wasn't going to do this at first, but this will also be great to have in my editor when upgrading modern Num Text versions to have feature parity with this version. It will and has already helped a lot to know the functionality of a given function call or property without having to look into it's actual implementation details to find out what it does. Going through it thoroughly a single time is much easier since you can use your documentation to understand it after that :)