Name Scope
Tried publishing this initial version of the module to npm, but it didn't like it's similarity in naming to the 'numtext' package, unfortunately.

At first I was going to rename the package to something else, but I already like using 'num-text', and it's fairly simple. I don't want a different unique name just for the sake of not colliding with other things.

Then I rediscovered the scope naming convention for npm packages, and I was considering my username as a scope, '@offroaders123'. It seemed a bit too long, and specific too. I paused for the night last night at that point.

When looking at this with new eyes in the morning, I thought about going with the '@ste' scope, but there are a few other grouped packages out there that fall under a commonly-named 'ste' acronym also, and I don't want to step on their toes by taking that scope from them if they end up wanting to use it down the line. '@stedit' doesn't appear to be colliding with anything else out there either, and it's already what I use for the app itself. It's easier to find everything too! Searching 'stedit' on Google shows the app, and likely eventually it's related npm packages too, once I publish them! I will probably go this way with 'menu-drop' too, making it '@stedit/menu-drop'. I also find it easier to type out than my whole username too. I'm not sure if I like username-scoped packages too much, unless it's something small. If it's something that might be used in multiple places, or in libraries outside your own, I feel like it doesn't have much room to grow from there. But maybe this is just a nitpick from my own perspective. My username is fairly long, so that's probably why. And I don't want to change it to something new at this point, it's also used to farly across my public reputation, haha.