CSS Modules Revert
This doesn't appear to be as close as it once seemed to be, now a few years back. Wow I can't believe how long it's been since I worked on this project really! Same with the reality of Menu-Drop and STE. I've been doing a heck ton bunch of refactoring and rebuilding, but I haven't added new features on their own in a while.

https://v8.dev/features/import-attributes
https://web.dev/articles/constructable-stylesheets