Complete Type Annotations
I think I've found out that doing this has really helped my brain with very quickly understanding how a piece of code works, and now I see why I've heard people say that in the past. That even happened today with the Squiggle demos! One of them was in TS, but it wasn't typed out (No implicit any seemed to be off because it was missing a tsconfig), and I found myself initially typing the function and API usages out, because then I knew how they were designed to fit together and how they 'should' work per se. So it's like a link between how the code works, and how it is intended to work, and I think types help tremendously with that.

I turned on dark mode for the demo because coding in the plane at night  with all the lights off, don't want to look like that 'flash grenade fortnite meme' guy 😂