Accessor Revert
The previous one was a possibility of how to set things up, but it's not completely there yet in terms of browser support, nor complete TypeScript/bundler support either. So I'm going to wait to try it later on, until it's more of a fully formed part of the spec/language.