ConnectedCallback Lifecycle
Moving back to the connectedCallback-style lifecycle, for the sake of me trying to conceptualize the lifecycles and how they should be managed in Web Components. I remember now that this wasn't an issue with using the constructor, because I am defining this component in a module script, which is always deferred to the end of the page. I want to see if I can allow this to work without needing that though, more for my understanding, once again, rather than for needing to support plain JS files themselves. Say like a non-module bundle, that could be helpful if someone wanted it to work like that out of the box.